schema-001 fixes

+++ DSL/Resql/analytics/GET/chat/list_chat_cif_long_waiting_time.sql
@@ -84,6 +84,6 @@ SELECT
     DATE_TRUNC(:period, created) AS time,
     COUNT(*) AS long_waiting_time
 FROM waiting_times
-WHERE waiting_time_seconds > :threshold_seconds
+WHERE waiting_time_seconds > :threshold_seconds::INTEGER
 GROUP BY time
 ORDER BY time;
